+++ commands/image.js
@@ -1,8 +1,54 @@
 const Discord = require("discord.js");
 const Command = require("../structures/command.js");
+const Jimp = require("jimp");
 const request = require("request");
 
 module.exports = [
+	class InvertCommand extends Command {
+		constructor() {
+			super({
+				name: "invert",
+				description: "Invert the colors of an image",
+				args: [
+					{
+						num: 1,
+						type: "image"
+					}
+				],
+				cooldown: {
+					time: 15000,
+					type: "channel"
+				},
+				perms: {
+					bot: ["ATTACH_FILES"],
+					user: [],
+					level: 0
+				},
+				usage: "invert <image URL>"
+			});
+		}
+		
+		async run(bot, message, args, flags) {
+			Jimp.read(args[0])
+			.then(img => {
+				img.invert().getBufferAsync(Jimp.MIME_PNG)
+				.then(imgToSend => {
+					message.channel.send({
+						files: [{
+							attachment: imgToSend,
+							name: "invert.png"
+						}]
+					})
+				})
+				.catch(err => {
+					message.channel.send("Failed to generate the image.")
+				})
+			})
+			.catch(err => {
+				message.channel.send("Failed to get image for that URL.")
+			})
+		}
+	},
 	class MemeCommand extends Command {
 		constructor() {
 			super({